feat: 비품 타입 등록 기능 구현
#BP-25

+++ src/main/java/com/example/beepoo/dto/ItemRequestDto.java
@@ -10,7 +10,7 @@ public class ItemRequestDto {
 
     private Integer seq;
     private String name;
-    private Integer typeCode;
+    private Long typeCode;
     private ItemStatusEnum status;
     private String serial;
     private String comment;
